+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - PR #2180 Fix issue with isort reordering `importorskip` below imports depending on them
 - PR #2187 fix to honor dtype when numpy arrays are passed to columnops.as_column
 - PR #2190 Fix issue in astype conversion of string column to 'str'
+- PR #2162 Setting is_unique and is_monotonic-related attributes
 
 
 # cuDF 0.8.0 (27 June 2019)